// src/app/share/enums/badge-colors.enum.ts
export enum BadgeColors {
  recordLocked = 'record-locked',
  specialHC = 'special-hc',
  maintPending = 'maint-pending',
  paperClaim = 'paper-claim',
  queue = 'queue',
  requestDenied = 'request-denied',
  requestApproved = 'request-approved',
  pendingApproval = 'pending-approval',
  memberMigrated = 'member-migrated'
}